Moved calls of CaseReferenceValidator to CallbackValidator
