+++ src/test/java/uk/gov/hmcts/reform/bulkscan/orchestrator/services/ccd/CallbackValidatorTest.java
@@ -3,23 +3,33 @@ package uk.gov.hmcts.reform.bulkscan.orchestrator.services.ccd;
 import io.vavr.control.Validation;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
 
 import java.util.Map;
 
 import static java.util.Collections.emptyMap;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.mock;
 
+@ExtendWith(MockitoExtension.class)
 class CallbackValidatorTest {
     private static final String SERVICE = "service";
     private static final String CASE_ID = "123";
     private static final String CASE_TYPE_ID = SERVICE + "_ExceptionRecord";
 
+    @Mock
+    private CaseReferenceValidator caseReferenceValidator;
+
     private CallbackValidator callbackValidator;
 
     @BeforeEach
     void setUp() {
-        callbackValidator = new CallbackValidator();
+        callbackValidator = new CallbackValidator(caseReferenceValidator);
     }
 
     @Test
@@ -118,5 +128,48 @@ class CallbackValidatorTest {
         // then
         assertThat(res.isValid()).isFalse();
         assertThat(res.getError()).isEqualTo("Internal Error: invalid jurisdiction supplied: null");
+    }
+
+    @Test
+    void hasTargetCaseReference_calls_caseReferenceValidator() {
+        // given
+        CaseDetails caseDetails = mock(CaseDetails.class);
+        Validation<String, String> validationRes = Validation.valid("caseRef");
+        given(caseReferenceValidator.validateTargetCaseReference(any(CaseDetails.class))).willReturn(validationRes);
+
+        // when
+        Validation<String,String> res = callbackValidator.hasTargetCaseReference(caseDetails);
+
+        // then
+        assertThat(res).isSameAs(validationRes);
+    }
+
+    @Test
+    void hasSearchCaseReference_calls_caseReferenceValidator() {
+        // given
+        CaseDetails caseDetails = mock(CaseDetails.class);
+        Validation<String, String> validationRes = Validation.valid("caseRef");
+        given(caseReferenceValidator.validateSearchCaseReferenceWithSearchType(any(CaseDetails.class)))
+                .willReturn(validationRes);
+
+        // when
+        Validation<String,String> res = callbackValidator.hasSearchCaseReference(caseDetails);
+
+        // then
+        assertThat(res).isSameAs(validationRes);
+    }
+
+    @Test
+    void hasSearchCaseReferenceType_calls_caseReferenceValidator() {
+        // given
+        CaseDetails caseDetails = mock(CaseDetails.class);
+        Validation<String, String> validationRes = Validation.valid("caseRef");
+        given(caseReferenceValidator.validateCaseReferenceType(any(CaseDetails.class))).willReturn(validationRes);
+
+        // when
+        Validation<String,String> res = callbackValidator.hasSearchCaseReferenceType(caseDetails);
+
+        // then
+        assertThat(res).isSameAs(validationRes);
     }
 }
